propagate away time.go
